+++ src/main/java/roomescape/payment/dto/TossPaymentErrorResponse.java
@@ -1,4 +1,4 @@
-package roomescape.payment.toss.dto;
+package roomescape.payment.dto;
 
 public record TossPaymentErrorResponse(String code, String message) {
 }
