--- src/main/java/roomescape/payment/toss/dto/TossPaymentRequest.java
@@ -1,14 +1,0 @@
-package roomescape.payment.toss.dto;
-
-import roomescape.reservation.controller.request.ReservePaymentRequest;
-
-public record TossPaymentRequest(String paymentKey, String orderId, Long amount) {
-
-    public static TossPaymentRequest from(ReservePaymentRequest request) {
-        return new TossPaymentRequest(
-                request.payment().paymentKey(),
-                request.payment().orderId(),
-                request.payment().amount()
-        );
-    }
-}
