+++ src/test/java/roomescape/payment/config/TestTossPaymentConfig.java
@@ -1,13 +1,12 @@
-package roomescape.payment.toss.config;
+package roomescape.payment.config;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import java.util.Base64;
-import okhttp3.mockwebserver.MockWebServer;
 import org.springframework.boot.test.context.TestConfiguration;
 import org.springframework.context.annotation.Bean;
 import org.springframework.http.client.SimpleClientHttpRequestFactory;
 import org.springframework.web.client.RestClient;
-import roomescape.payment.toss.interceptor.TossPaymentResponseInterceptor;
+import roomescape.payment.interceptor.TossPaymentResponseInterceptor;
 
 @TestConfiguration
 public class TestTossPaymentConfig {
